+++ packages/express-layer/index.js
@@ -30,6 +30,7 @@ module.exports = engine => {
                 ? context.request.body
                 : JSON.stringify(context.request.body),
             headers: context.headers,
+            followRedirect: false,
           },
           (error, response, body) => {
             context.res = response;